add logic for verify master hostnetworkpod
